Tweaks to pull request for fixing genre scraping.
- Remove Shounen from integration test, as Eureka Seven no longer
has that genre on MAL.
- Remove "#pragma warning disable" and "#pragma warning restore"
from Eureka Seven HTML in unit test project. The HTML should be
exactly as it is from MAL.
- Remove HTML Agility Pack from unit test project. The unit
test project does not need to reference it.
- Remove failing unit test demonstrating old code failure. I wasn't
clear in my comments on the pull request that there should be unit
tests that fail with the old code and succeed with the new code.
- Remove "Returns DateTime?" doc comment from UncertainDate.ToDateTime(),
it tells you nothing that the method signature doesn't.
- Use foreach loop instead of for loop to iterate over matching html
nodes.
- Use .Attributes["href"] instead of .Attributes.First() to be resilient
against attributes being added before href.
- Match link URL against /genre/<genreID> instead of just any numbers
in the URL to be resilient against other numbers being added to the URL.
- Use InnerText instead of InnerHTML to be resilient against HTML being
added inside the <a> tag.