Addressed almost all of your comments.
Regarding the response from the update: there are two ways to break the XML - in the first scenario it returns a message for invalid XML and in the other it a MySQL error and status OK ("Updated"). Unfortunately the response does not contain the MySQL information bit. On the other hand I don't know how you can break the XML with the current configuration.

When credentials are not correct the already existing code takes care of it and returns a 401.