Added some tests.
Added integration tests for the update functionality. I apologize for the poor execution but my experience with tests is very limtied.

There is some strange occurance where some of the tests fail depending on the name of the Initialize function. I guess the name somehow is responsible for the order of execution but I could not find information about it.